requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
pyodbc